Relazione e disabilitazione dropdownlist in Consuntivazione

+++ tcfWeb/client/src/app/consuntivazione/userDetail/monthGrid/monthGrid.component.ts
@@ -7,15 +7,16 @@ import { User } from '../../../model/user';
 import { ConsuntivazioneService } from '../../../service/consuntivazione.service';
 import { SystemService } from '../../../service/system.service';
 import { AttivitaService } from '../../../service/attivita.service';
-import { SelectItem } from 'primeng/primeng';
+import { SelectItem, ConfirmationService } from 'primeng/primeng';
 import { ClienteService } from '../../../service/cliente.service';
 import { MeseConsuntivoService } from '../../../service/meseConsuntivo.service';
+import { AmbitoService } from '../../../service/ambito.service';
 
 @Component({
   selector: 'month-grid',
   templateUrl: './monthGrid.component.html',
   styleUrls: ['./monthGrid.component.css'],
-  providers: [FormBuilder]
+  providers: [FormBuilder, AmbitoService, ConfirmationService]
 })
 
 export class MonthGridComponent implements OnChanges {
@@ -57,7 +58,9 @@ export class MonthGridComponent implements OnChanges {
     private clienteService: ClienteService,
     private meseConsuntivoService: MeseConsuntivoService,
     private systemService: SystemService,
-    private formBuilder: FormBuilder
+    private ambitoService: AmbitoService,
+    private formBuilder: FormBuilder,
+    private confirmationService: ConfirmationService
   ) {
 
     this.consuntivoForm = this.formBuilder.group({
@@ -74,11 +77,8 @@ export class MonthGridComponent implements OnChanges {
 
     //POPOLAMENTO LISTE
     this.lst_clienti = new Array<SelectItem>();
-    this.clienteService.getClienti().subscribe(clienti => {
-      clienti.forEach(cliente => {
-        this.lst_clienti.push({ label: cliente.nome_cliente, value: cliente._id });
-      });
-    });
+    
+    
 
     this.lst_attivita = new Array<SelectItem>();
     this.attivitaService.getAttivita().subscribe(attivitas => {
@@ -94,10 +94,6 @@ export class MonthGridComponent implements OnChanges {
 
     this.systemService.getAree().subscribe(domain => {
       this.lst_aree = domain;
-    });
-
-    this.systemService.getAmbiti().subscribe(domain => {
-      this.lst_ambiti = domain;
     });
 
 
@@ -254,9 +250,19 @@ export class MonthGridComponent implements OnChanges {
   //NEW ROW
   private showDialogToAdd() {
     //Inizializzazione dei parametri di input
+    this.lst_clienti = [];
+    this.clienteService.getClienti().subscribe(clienti => {
+      clienti.forEach(clienteAll => {
+        this.userSelected.clienti.forEach(clienteUser => {
+          if(clienteAll._id == clienteUser.id_cliente)
+          this.lst_clienti.push({ label: clienteAll.nome_cliente, value: clienteAll._id});
+        });
+      });
+    });
     this.resetConsuntivo(this.newRowConsuntivo);
     this.newRowConsuntivo.id_utente = this.userSelected._id;
-
+    this.consuntivoForm.reset();
+    this.formSubmitted = false;
     this.displayDialog = true;
 
   }
@@ -383,16 +389,23 @@ export class MonthGridComponent implements OnChanges {
     delCriteria.id_attivita = r.id_attivita;
     delCriteria.id_tipo_deliverable = r.id_tipo_deliverable;
 
-    this.consuntivazioneService
-      //.deleteConsuntivi(r.id_user, r.id_macro_area, r.id_ambito, r.id_attivita, r.id_tipo_deliverable)
-      .deleteConsuntivi(delCriteria)
-      .subscribe(msg => {
-        this.consuntivi.splice(i, 1);
-        this.consuntivi = Object.create(this.consuntivi); //deepcopy    
-        alert("deleted " + msg);
-      },
-      err => alert(err)
-      );
+    this.confirmationService.confirm({
+      message: "Sei sicuro di voler eliminare l'attività '" + r.nome_attivita + "' ?",
+      header: 'Elimina attività',
+      icon: 'fa fa-trash',
+      accept: () => {
+
+        this.consuntivazioneService
+          //.deleteConsuntivi(r.id_user, r.id_macro_area, r.id_ambito, r.id_attivita, r.id_tipo_deliverable)
+          .deleteConsuntivi(delCriteria)
+          .subscribe(msg => {
+            this.consuntivi.splice(i, 1);
+            this.consuntivi = Object.create(this.consuntivi); //deepcopy    
+          },
+          err => alert(err)
+          );
+      }
+    });
 
   }
 
@@ -439,10 +452,54 @@ export class MonthGridComponent implements OnChanges {
     return form.valid;
   }
 
+  private selectFromCliente(componentname) {
+    var selCriteria;
+    selCriteria = new Object();
+    selCriteria.id_cliente = this.newRowConsuntivo.id_cliente;
+
+    switch (componentname) {
+      case 'attivita':
+        this.lst_attivita = [];
+        this.attivitaService.getAttivitaByCliente(selCriteria).subscribe(commesse => {
+          commesse.forEach(element => {
+            this.lst_attivita.push({ label: element.nome_attivita, value: element.codice_attivita });
+          });
+        })
+        break;
+      case 'ambito':
+        this.lst_ambiti = [];
+        this.ambitoService.getAmbitoByCliente(selCriteria).subscribe(commesse => {
+          commesse.forEach(element => {
+            this.lst_ambiti.push({ label: element.nome_ambito, value: element._id });
+          });
+        })
+        break;
+    }
+  }
+
+  private isDisabled(componentName): boolean {
+    var disabled = false;
+
+    switch (componentName) {
+      case 'ambito': disabled = this.newRowConsuntivo.id_cliente == null;
+        break;
+      case 'macro_area': disabled = this.newRowConsuntivo.id_cliente == null;
+        break;
+      case 'attivita': disabled = this.newRowConsuntivo.id_cliente == null || this.newRowConsuntivo.id_ambito == null || this.newRowConsuntivo.id_macro_area == null;
+        break;
+      case 'deliverable': disabled = this.newRowConsuntivo.id_cliente == null || this.newRowConsuntivo.id_ambito == null || this.newRowConsuntivo.id_macro_area == null || this.newRowConsuntivo.id_attivita == null;
+        break;
+    }
+
+    return disabled;
+  }
+
   private isValid(componentName: string) {
     if ((this.consuntivoForm.get(componentName).touched || this.formSubmitted) && this.consuntivoForm.get(componentName).errors)
       return "#a94442";
     else
       return "#898989"; //#d6d6d6
   }
+
+
 }
